Fixed small bugs

+++ src/main/database_scripts/create/table/create_dish_table.sql
@@ -1,11 +1,11 @@
 CREATE TABLE dish
 (
-  ID          int         NOT NULL AUTO_INCREMENT,
-  Name        varchar(90) NOT NULL,
-  Description varchar(1000),
-  Cost        decimal     NOT NULL,
-  ImageHref   varchar(90),
-  IsInMenu    boolean     NOT NULL,
+    ID          int            NOT NULL AUTO_INCREMENT,
+    Name        varchar(90)    NOT NULL,
+    Description varchar(1000),
+    Cost        decimal(10, 2) NOT NULL,
+    ImageHref   varchar(90),
+    IsInMenu    boolean        NOT NULL,
 
-  PRIMARY KEY (ID)
+    PRIMARY KEY (ID)
 );
